requirements.txt
# packages in environment at D:\Professional\Master\MTH003 - Cac He Co So Du Lieu Nang Cao\Final Project\Source\Implementation1\venv:
#
# Name                    Version                   Build  Channel
ca-certificates           2021.4.13            haa95532_1  
certifi                   2020.12.5        py37haa95532_0  
icu                       58.2                 ha925a31_3  
jpeg                      9b                   hb83a4c4_2  
libpng                    1.6.37               h2a8f88b_0  
libprotobuf               3.6.0                h1a1b453_0  
mysql-connector-c         6.1.11               h33f27b4_1  
mysql-connector-python    8.0.18           py37h13ed8b8_1  
neo4j-python-driver       4.2.1            py37h2bbff1b_0  
neotime                   1.7.4                    py37_0  
openssl                   1.1.1k               h2bbff1b_0  
pip                       21.1.1           py37haa95532_0  
protobuf                  3.6.0            py37he025d50_0  
pymongo                   3.11.4           py37hd77b12b_0  
pyqt                      5.9.2            py37h6538335_2  
python                    3.7.10               h6244533_0  
pytz                      2021.1             pyhd3eb1b0_0  
qt                        5.9.7            vc14h73c81de_0  
qtawesome                 1.0.2              pyhd3eb1b0_0  
qtpy                      1.9.0                      py_0  
redis-py                  3.5.3                      py_0  
setuptools                52.0.0           py37haa95532_0  
sip                       4.19.8           py37h6538335_0  
six                       1.15.0           py37haa95532_0  
sqlite                    3.35.4               h2bbff1b_0  
vc                        14.2                 h21ff451_1  
vs2015_runtime            14.27.29016          h5e58377_2  
wheel                     0.36.2             pyhd3eb1b0_0  
wincertstore              0.2                      py37_0  
zlib                      1.2.11               h62dcd97_4  
